// src/app/model/playerstats.ts
import { Data } from './data';

export class Playerstats {
    data: Data = new Data();
    creditsLeft: string = "";
    pid: string = "";
    country: string = "";
    profile: string = "";
    imageURL: string = "";
    battingStyle: string = "";
    bowlingStyle: string = "";
    majorTeams: string = "";
    currentAge: string = "";
    born: string = "";
    fullName: string = "";
    name: string = "";
    playingRole: string = "";
}